+++ client/src/components/ResumeAccordion.js
@@ -19,8 +19,6 @@ import {
 
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import DeleteIcon from "@mui/icons-material/Delete";
-import jsPDF from "jspdf";
-import html2canvas from "html2canvas";
 
 import { useSelector, useDispatch } from "react-redux";
 import {
@@ -141,22 +139,7 @@ const ResumeAccordion = () => {
   const handleDeleteProject = (index) => {
     dispatch(deleteProject(index));
   };
-  const rightPartRef = useRef(null);
-  const handleDownloadPDF = () => {
-    const input = rightPartRef.current;
 
-    // Use html2canvas to capture the content as an image
-    html2canvas(input).then((canvas) => {
-      const imgData = canvas.toDataURL("image/png");
-
-      // Use jsPDF to create a PDF document
-      const pdf = new jsPDF();
-      pdf.addImage(imgData, "JPEG", 0, 0);
-
-      // Save the PDF with a specific name (e.g., "resume.pdf")
-      pdf.save("resume.pdf");
-    });
-  };
   const [expandedAccordion, setExpandedAccordion] = useState(null);
 
   const handleAccordionChange = (panel) => (event, isExpanded) => {
@@ -667,17 +650,6 @@ const ResumeAccordion = () => {
           ))}
         </AccordionDetails>
       </Accordion>
-      {/* Button to download the right part as PDF */}
-      <Grid item xs={4}>
-        <Button
-          variant="contained"
-          color="primary"
-          onClick={handleDownloadPDF}
-          style={{ marginTop: "16px" }}
-        >
-          Download
-        </Button>
-      </Grid>
     </Grid>
   );
 };
